Allow usertags as aliases
This should allow +1 @andrzej once the alias is made.
For now we only allow aliases for user tags. This limits the scope of
problems. Aliasing any name requires merging existing points.

Old migrations required to be inherited from specific versions of
ActiveRecord::Migration - thus the invasive change.

For now the implementation assumes a new model - Alias.
Hopefully, the next step would be to refactor it to commands/events.